requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.4
django-allauth==0.62.0
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-filter==24.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pillow==10.3.0
python-dotenv==1.0.1
pywin32==306
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.5.0
tzdata==2024.1
uploads==1.1
uritemplate==4.1.1

